Foo
Because paths can contain backslashes, searching for the first backslash
starting on the right for detecting namespaces yields the wrong result.
Furthermore, because the namespace of the class is not actually in the
name (only the interface or parent class is) the namespace can be wrong.

This fix consists of two parts:
 - Fix how a prefix is chosen such that the correct namespace is included in
   the name.
 - If ZEND_ACC_ANON_CLASS is set, check the position of @ and search for the
   backslash before that.

While the issue seems Windows-specific, I could even reproduce this on Linux
because Linux allows directories to contain backslashes.